refactor(headless): split 772-line god file into events, UI, and context modules
